Extract persistence utilities to persistence_utils.rs
Move response persistence code from conversations/handlers.rs to a
new persistence_utils.rs file to enable reuse across routers.

Changes:
- Create routers/persistence_utils.rs with shared persistence functions
- Update conversations/handlers.rs to only contain API handlers
- Update imports in openai/router.rs, openai/streaming.rs, grpc/utils.rs
