+++ sgl-model-gateway/src/routers/mod.rs
@@ -29,6 +29,7 @@ pub mod http;
 pub mod mcp_utils;
 pub mod openai;
 pub mod parse;
+pub mod persistence_utils;
 pub mod router_manager;
 pub mod tokenize;
